+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+**1.18.0** 
+* Initial release of Q search bar embedding support
+  * Added new set of options for QSearchBar
+
 **1.17.1** 
 * Minor bug fixes
   * Eliminate usage of eval in non-minified js file